requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
fastapi
uvicorn
python-dotenv
langchain-groq
pydantic
watchdog
pypdf
rank_bm25
